+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Fix bug in Negate expression function.
+
 * Issue **#1995** : Add help info to the expression functions drop down menu.
 
 * Issue **#1911** : Add a drop down menu for picking index fields in the expression editor.